refactor(detectors): move resolve_or_train_detector into vtscore (Theme A)
The cold-path detector model resolution/training logic lived inline in the
scoring route handler despite having no Flask or request-context dependency.
Move it to vtscore.detectors.model_loading.resolve_or_train_detector so the
library tier owns it and the route is request<->library glue.

First slice of the code-structure review's Theme A (vtscore/vtsearch
boundary). See docs/plans/code-structure-review.md.
