+++ vtscore/detectors/dataset_sync.py
@@ -236,7 +236,7 @@ def ensure_detector_model_matches_active_embedder() -> None:
     that the on-disk labelset doesn't carry.
 
     The scoring fast-paths
-    (:func:`vtsearch.routes.detectors.scoring._resolve_or_train_detector`
+    (:func:`vtscore.detectors.model_loading.resolve_or_train_detector`
     and the find dispatcher) defensively repeat the check per-detector,
     since Auto-Find / multi-dataset Find iterate detectors that aren't the
     active one.
